requirements.txt
python-fasthtml==0.2.1